Re-introduce the Factory class
