+++ composer.json
@@ -31,11 +31,12 @@
     },
     "require": {
         "php": ">=7.1.3",
-        "ext-mbstring" : "*",
         "ext-fileinfo": "*",
-        "psr/http-message": "^1.0",
+        "ext-mbstring": "*",
+        "league/uri-interfaces": "dev-develop",
         "league/uri-parser": "dev-develop",
-        "league/uri-interfaces": "dev-develop"
+        "psr/http-factory": "^1.0",
+        "psr/http-message": "^1.0"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.0",
